requirements.txt
# ── Core Framework ──────────────────────────────────────────────────
streamlit>=1.35.0

# ── Data ────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0
python-dateutil>=2.9.0

# ── Visualización ───────────────────────────────────────────────────
plotly>=5.22.0
matplotlib>=3.8.0

# ── PDF / Documentos ────────────────────────────────────────────────
fpdf2>=2.7.9
reportlab>=4.2.0
PyMuPDF>=1.24.0
python-docx>=1.1.0

# ── Imágenes / CV ───────────────────────────────────────────────────
Pillow>=10.3.0
opencv-python-headless>=4.9.0.80
qrcode[pil]>=7.4.2

# ── Seguridad / Auth ────────────────────────────────────────────────
bcrypt>=4.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0

# ── AI / Google ─────────────────────────────────────────────────────
google-generativeai>=0.7.0

# ── Audio / TTS ─────────────────────────────────────────────────────
gTTS>=2.5.0

# ── HTTP ────────────────────────────────────────────────────────────
requests>=2.31.0
